requirements.txt
absl-py==1.3.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
configer==1.3.1
configparser==5.3.0
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
filelock==3.8.0
fonttools==4.37.3
freetype-py==2.3.0
future==0.18.2
gdown==4.5.1
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gshell==5.6.1
human-body-prior==0.8.5.0
idna==3.4
imageio==2.22.0
importlib-metadata==4.12.0
ipython==7.34.0
jedi==0.18.1
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
natsort==8.2.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.20.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.27
Pygments==2.13.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.39
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smplx==0.1.28
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch
torchfile
torchgeometry
torchvision
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transforms3d==0.4.1
trimesh==3.15.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.12
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.8.1
dadaptation==3.1
